requirements.txt
python-gitlab==3.13.0
argparse==1.4.0
